+++ lib/board.ex
@@ -30,14 +30,14 @@ defmodule Board do
 
   def is_there_a_winner?(grid) do
     grid
-    |> all_winning_lines
-    |> Enum.any?(fn line -> winning_line?(line) end)
+    |> Line.all_winning_lines
+    |> Enum.any?(fn line -> Line.winning_line?(line) end)
   end
 
   def winning_move(grid) do
     grid
-    |> all_winning_lines
-    |> Enum.find(fn line -> winning_line?(line) end)
+    |> Line.all_winning_lines
+    |> Enum.find(fn line -> Line.winning_line?(line) end)
     |> Enum.uniq
     |> Enum.at(0)
   end
@@ -48,40 +48,7 @@ defmodule Board do
     |> Enum.random
   end
 
-  def all_winning_lines(grid) do
-    rows(grid) ++ columns(grid) ++ diagonals(grid)
-  end
-
   defp create_grid(number) do
     Enum.to_list 0..number
-  end
-
-  defp winning_line?(line) do
-    length(Enum.uniq(line)) == 1
-  end
-
-  defp rows(grid) do
-    grid
-    |> Enum.chunk_every(size(grid))
-  end
-
-  defp columns(grid) do
-    grid
-    |> rows
-    |> List.zip
-    |> Enum.map(&Tuple.to_list/1)
-  end
-
-  defp diagonals(grid) do
-    [diagonal_line([], grid, 0, size(grid) + 1), diagonal_line([], grid, size(grid) - 1, size(grid) - 1)]
-  end
-
-  defp diagonal_line(line, grid, current_index, increase_index_by) do
-    if length(line) == size(grid) do
-      line
-    else
-      line ++ [Enum.at(grid, current_index)]
-      |> diagonal_line(grid, current_index + increase_index_by, increase_index_by)
-    end
   end
 end
